feat: wip - clean up, continue with auth scaffolding
