+++ backend/go.mod
@@ -3,15 +3,13 @@ module github.com/expatx/expatx/backend
 go 1.20
 
 require (
-	github.com/jinzhu/gorm v1.9.16
-	github.com/pkg/errors v0.9.1
+	github.com/golang-jwt/jwt v3.2.2+incompatible
 	google.golang.org/grpc v1.56.0
 	google.golang.org/protobuf v1.30.0
 )
 
 require (
 	github.com/golang/protobuf v1.5.3 // indirect
-	github.com/jinzhu/inflection v1.0.0 // indirect
 	golang.org/x/net v0.9.0 // indirect
 	golang.org/x/sys v0.7.0 // indirect
 	golang.org/x/text v0.9.0 // indirect
