+++ backend/pkg/db/db.go
@@ -1,0 +1,2 @@
+// Package db provides database operations.
+package db
