+++ backend/proto/v1/auth/auth.proto
@@ -12,21 +12,41 @@ message SignupRequest {
 }
 
 message SignupResponse {
-    string token = 1;
-    string error = 2;
-    user.v1.User user = 3;
+  string token = 1;
+  string error = 2;
+  user.v1.User user = 3;
 
-    enum Status {
-        OK = 0;
-        INVALID_EMAIL = 1;
-        INVALID_PASSWORD = 2;
-        EMAIL_ALREADY_EXISTS = 3;
-        UNKNOWN = 4;
-    }
+  enum Status {
+    OK = 0;
+    INVALID_EMAIL = 1;
+    INVALID_PASSWORD = 2;
+    EMAIL_ALREADY_EXISTS = 3;
+    UNKNOWN = 4;
+  }
 
-    Status status = 4;
+  Status status = 4;
+}
+
+message LoginRequest {
+  string username = 1;
+  string password = 2;
+}
+
+message LoginResponse {
+  string token = 1;
+  string error = 2;
+  user.v1.User user = 3;
+
+  enum Status {
+    OK = 0;
+    INVALID_CREDENTIALS = 1;
+    UNKNOWN = 2;
+  }
+
+  Status status = 4;
 }
 
 service AuthService {
   rpc Signup(SignupRequest) returns (SignupResponse);
+  rpc Login(LoginRequest) returns (LoginResponse);
 }
