+++ backend/proto/v1/user/user.proto
@@ -8,11 +8,13 @@ import "google/protobuf/timestamp.proto";
 
 message User {
   string id = 1;
-  string name = 2;
-  string email = 3;
-  string password = 4;
-  google.protobuf.Timestamp created_at = 5;
-  google.protobuf.Timestamp updated_at = 6;
+  string first_name = 2;
+  string last_name = 3;
+  string username = 4;
+  string email = 5;
+  string password = 6;
+  google.protobuf.Timestamp created_at = 7;
+  google.protobuf.Timestamp updated_at = 8;
 }
 
 message CreateUserRequest {
